Created CleanupTask for cleaning up md directories.
This could be VERY dangerous if it isn't working as expected, so I have
not yet added this to any workflows generated by this library. For
workflows it is added to, it should probably be an opt-in kwarg to
include it.

It will delete 'files' in the fw_spec, which in this context
would be defined by BeaconTask. BeaconTask currently only puts the
'SCRATCHDIR' environment variable in this list of files, so that should
be the only thing deleted. It WOULD be catastrophic if somebody made
SCRATCHDIR a directory they didn't want to be deleted in full after a
run (like their home directory...).