Now we run md out of rundir in scratch named by uuid.
To facilitate running multiple MD tasks in series or in parallel for a
single job, we don't want all files transferred to the exact same
scratch directory. So, instead, just like in staging we transfer files
to a directory on scratch with the Sim's uuid as the directory name.

This requires the run_md.sh script to read the run directory from stdin.
We need to add a template script for this to show what must be there for
all this to work well.