Bugfixes, plus now using a MkRunDirTask for building scratch dir.
The trouble with transferring files to a directory inside of the
scratchdir set up by our template queue script is that this uuid-named
directory doesn't exist yet, and the FileTransferTask will not work as
written. We could use the 'copytree' mode of FileTransferTask to get
around this, but this will fail for a directory that already exists,
which would happen in the case where the next iteration of a Sim is run
by the same queue session.

To get around this we just added a dedicated Task for making the
directory, and it is perfectly fine with the directory already existing.