CleanupTask much safer, no longer recursive.
At worst, it will delete all the files in a directory and then try to
delete the directory itself. This will fail if it still has things in
it, such as other directories. Since we only use it to delete things
that the BeaconTask pointed out for Pulling, we should be safe here.

Regardless, it's opt-in for ``make_md_workflow``.